+++ tests/test_api.py
@@ -1,5 +1,6 @@
 import pytest
 
+from api_tabular import config
 from api_tabular.utils import external_url
 
 from .conftest import DATE, PGREST_ENDPOINT, RESOURCE_ID, TABLES_INDEX_PATTERN
@@ -108,7 +109,7 @@ async def test_api_resource_data_with_args_case(client, rmock):
     args = "COLUM_NAME__EXACT=BIDULE&page=1"
     rmock.get(TABLES_INDEX_PATTERN, payload=[{"__id": 1, "id": "test-id", "parsing_table": "xxx"}])
     rmock.get(
-        f"{PGREST_ENDPOINT}/xxx?COLUM_NAME=eq.BIDULE&limit=20&order=__id.asc",
+        f'{PGREST_ENDPOINT}/xxx?"COLUM_NAME"=eq.BIDULE&limit=20&order=__id.asc',
         payload={"such": "data"},
         headers={"Content-Range": "0-10/10"},
     )
@@ -152,7 +153,7 @@ async def test_api_resource_data_with_page_size_error(client, rmock):
         "errors": [
             {
                 "code": None,
-                "detail": "Page size exceeds allowed maximum",
+                "detail": f"Page size exceeds allowed maximum: {config.PAGE_SIZE_MAX}",
                 "title": "Invalid query string",
             }
         ]
@@ -179,7 +180,7 @@ async def test_api_resource_data_table_error(client, rmock):
 async def test_api_percent_encoding_arabic(client, rmock):
     rmock.get(TABLES_INDEX_PATTERN, payload=[{"__id": 1, "id": "test-id", "parsing_table": "xxx"}])
     rmock.get(
-        f"{PGREST_ENDPOINT}/xxx?%D9%85%D9%88%D8%A7%D8%B1%D8%AF=eq.%D9%85%D9%88%D8%A7%D8%B1%D8%AF&limit=20&order=__id.asc",  # noqa
+        f'{PGREST_ENDPOINT}/xxx?"%D9%85%D9%88%D8%A7%D8%B1%D8%AF"=eq.%D9%85%D9%88%D8%A7%D8%B1%D8%AF&limit=20&order=__id.asc',  # noqa
         status=200,
         payload={"such": "data"},
         headers={"Content-Range": "0-10/10"},